can't use rootResolvers in apoolo :/ may never use apollo again
